authorization done

+++ app/Http/Controllers/V1/ImageManipulationController.php
@@ -10,17 +10,19 @@ use Illuminate\Support\Str;
 use Illuminate\Support\Facades\File;
 use Intervention\Image\Facades\Image;
 use App\Http\Resources\V1\ImageManipulationResource;
+use Illuminate\Http\Request;
 
 class ImageManipulationController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\Resources\Json\AnonymousResourceCollection
      */
-    public function index()
+    public function index(Request $request)
     {
-        return ImageManipulationResource::collection(ImageManipulation::paginate());
+        // return ImageManipulationResource::collection(ImageManipulation::paginate());
+        return ImageManipulationResource::collection(ImageManipulation::where('user_id', $request->user()->id)->paginate());
     }
 
     /**
@@ -41,11 +43,16 @@ class ImageManipulationController extends Controller
         $data = [
             'type' => ImageManipulation::TYPE_RESIZE,
             'data' => json_encode($all),
-            'user_id' => null
+            'user_id' => $request->user()->id
         ];
 
         if(isset($all['album_id'])) {
-            // TODO
+            $album = Album::find($all['album_id']);
+            
+            if($request->user()->id != $album->user_id) {
+                return abort(403, 'Unauthorized');
+            }
+            
             $data['album_id'] = $all['album_id'];
         }
 
@@ -98,8 +105,12 @@ class ImageManipulationController extends Controller
      * @param  \App\Models\ImageManipulation  $image
      * @return \Illuminate\Http\Response
      */
-    public function show(ImageManipulation $image)
+    public function show(Request $request, ImageManipulation $image)
     {
+        if($request->user()->id != $image->user_id) {
+            return abort(403, 'Unauthorized');
+        }
+
         return new ImageManipulationResource($image);
     }
 
@@ -109,8 +120,12 @@ class ImageManipulationController extends Controller
      * @param  \App\Models\Album $album
      * @return \Illuminate\Http\Response
      */
-    public function byAlbum(Album $album)
+    public function byAlbum(Request $request, Album $album)
     {
+        if($request->user()->id != $album->user_id) {
+            return abort(403, 'Unauthorized');
+        }
+
         $where = ['album_id' => $album->id];
         return ImageManipulationResource::collection(ImageManipulation::where($where)->paginate());
     }
@@ -121,8 +136,12 @@ class ImageManipulationController extends Controller
      * @param  \App\Models\ImageManipulation  $image
      * @return \Illuminate\Http\Response
      */
-    public function destroy(ImageManipulation $image)
+    public function destroy(Request $request, ImageManipulation $image)
     {
+        if($request->user()->id != $image->user_id) {
+            return abort(403, 'Unauthorized');
+        }
+
         $image->delete();
 
         return response('', 204);
